move the SupportabilityMetrics class and make package-private
